move errors into generate_message
